Crate pnmixerlib

Reexports

pub extern crate flexi_logger;

pub extern crate log;

pub extern crate error_chain;

pub extern crate serde_derive;

pub extern crate toml;

pub extern crate serde;

pub extern crate alsa;

pub extern crate alsa_sys;

pub extern crate ffi;

pub extern crate gdk;

pub extern crate gdk_pixbuf;

pub extern crate gdk_pixbuf_sys;

pub extern crate gdk_sys;

pub extern crate gio;

pub extern crate glib;

pub extern crate glib_sys;

pub extern crate gobject_sys;

pub extern crate gtk;

pub extern crate gtk_sys;

pub extern crate libc;

pub extern crate png;

pub extern crate which;

pub extern crate xdg;

Modules

alsa_card
Alsa audio subsystem.

app_state
Global application state.

audio
High-level audio subsystem.

errors

glade_helpers

prefs
The preferences subsystem.

support_alsa
Alsa audio helper functions.

support_audio
Helper functions of the audio subsystem.

support_cmd
Helper functions for invoking system commands.

support_ui
Helper functions for the UI, mostly pixbuf functions.

ui_entry
Global GUI state.

ui_popup_menu
The popup menu subsystem when the user right-clicks on the tray icon.

ui_popup_window
The popup window subsystem when the user left-clicks on the tray icon.

ui_prefs_dialog
The preferences window subsystem, when the user clicks the "Preferences" menu item on the popup menu.

ui_tray_icon
The tray icon subsystem.

Macros

create_builder_item

pixbuf_new_from_png
Create a pixbuf from the given PNG file. Includes the file as bytes in the binary and decodes it.

try_e

try_er

try_r

try_w

try_wr