Module pnmixerlib::gdk_x11

Glue code between gdk and x11, allowing some gdk_x11_* functions.

This is not a complete binding, but just provides what we need in a reasonable way.

Functions

gdk_x11_get_default_root_xwindow
Gets the root window of the default screen (see gdk_x11_get_default_screen()).

gdk_x11_get_default_xdisplay
Gets the default GTK+ display.

gdk_x11_window_foreign_new_for_display
Wraps a native window in a GdkWindow. The function will try to look up the window using gdk_x11_window_lookup_for_display() first. If it does not find it there, it will create a new window.