Function pnmixerlib::gdk_x11::gdk_x11_window_foreign_new_for_display

pub fn gdk_x11_window_foreign_new_for_display(
    gdk_display: &mut Display, 
    xwindow: Window
) -> Option<Window>

Wraps a native window in a GdkWindow. The function will try to look up the window using gdk_x11_window_lookup_for_display() first. If it does not find it there, it will create a new window.

This may fail if the window has been destroyed. If the window was already known to GDK, a new reference to the existing GdkWindow is returned.

display

the GdkDisplay where the window handle comes from.

window

an Xlib Window

Returns

a GdkWindow wrapper for the native window, or None if the window has been destroyed. The wrapper will be newly created, if one doesn’t exist already.