Struct pnmixerlib::prefs::Prefs

pub struct Prefs {
    pub device_prefs: DevicePrefs,
    pub view_prefs: ViewPrefs,
    pub behavior_prefs: BehaviorPrefs,
}

Main preferences struct, holding all sub-preferences.

Fields
device_prefs: DevicePrefs view_prefs: ViewPrefs behavior_prefs: BehaviorPrefs
Methods

impl Prefs

fn new() -> Result<Prefs>

fn reload_config(&mut self) -> Result<()>

Reload the current preferences from the config file.

fn store_config(&self) -> Result<()>

Store the current preferences to the config file.

fn to_str(&self) -> String

Conver the current preferences to a viewable String.

fn get_avail_vol_control_cmd(&self) -> Option<String>

Get an available volume control command, which must exist in $PATH. Tries hard to fine one, starting with the given preference setting and falling back to the VOL_CONTROL_COMMANDS slice.