Module pnmixerlib::support_alsa

Alsa audio helper functions.

This mod wraps around a few low-level alsa functions and abstracts out the details we don't care about.

Functions

alsa_card_has_playable_selem
Check whether the given alsa card as a playable Selem.

get_alsa_card_by_id
Get an alsa card corresponding to the given ID.

get_alsa_card_by_name
Get an alsa card by the given name.

get_alsa_cards
Get all available alsa cards.

get_default_alsa_card
Get the default alsa card. This is the one with the ID 0.

get_first_playable_alsa_card
Get the first playable alsa card.

get_first_playable_selem
Get the first playable Selem.

get_mixer
Get the Mixer for the given alsa card.

get_playable_alsa_card_names
Get the names of all playable alsa cards.

get_playable_selem_by_name
Get a playable Selem by the given name.

get_playable_selem_names
Get the names from all playable Selems.

get_playable_selems
Get all playable Selems.

get_selem
Get the Selem from the given Elem.

selem_is_playable
Check whether the given Selem is playable.