Module pnmixerlib::ui_popup_window

The popup window subsystem when the user left-clicks on the tray icon.

This shows the manipulatable volume slider with the current volume and the mute checkbox.

Structs

PopupWindow
The main struct for the popup window, holding all relevant sub-widgets and some mutable state.

Functions

init_popup_window
Initialize the popup window subsystem.

set_slider
Set the volume slider to the given value.