Struct pnmixerlib::ui_popup_window::PopupWindow

pub struct PopupWindow {
    pub popup_window: Window,
    pub vol_scale_adj: Adjustment,
    pub vol_scale: Scale,
    pub mute_check: CheckButton,
    pub mixer_button: Button,
    // some fields omitted
}

The main struct for the popup window, holding all relevant sub-widgets and some mutable state.

Fields
popup_window: Window

The main window for the popup window widget.

vol_scale_adj: Adjustment

The volume scale adjustment.

vol_scale: Scale

The volume scale.

mute_check: CheckButton

The mute check button.

mixer_button: Button

The button to start the external mixer.

Methods

impl PopupWindow

fn new(builder: Builder) -> PopupWindow

Constructor.

fn update(&self, audio: &Audio) -> Result<()>

Update the popup window state, including the slider and the mute checkbutton.

fn update_mute_check(&self, audio: &Audio)

Update the mute checkbutton.