Module pnmixerlib::ui_tray_icon

The tray icon subsystem.

This manages the tray icon Pixbuf as well as the callbacks on left and right-click.

Structs

AudioPix
The actual tray icon Pixbuf, which depends on the current volume level.

TrayIcon
The tray icon struct, describing the complete visual state.

VolMeter
The volume meter, describes by its colors, offset and width/row properties.

Functions

init_tray_icon
Initialize the tray icon subsystem.