Struct pnmixerlib::ui_tray_icon::TrayIcon

pub struct TrayIcon {
    pub volmeter: RefCell<Option<VolMeter>>,
    pub audio_pix: RefCell<AudioPix>,
    pub status_icon: StatusIcon,
    pub icon_size: Cell<i32>,
    // some fields omitted
}

The tray icon struct, describing the complete visual state.

Fields
volmeter: RefCell<Option<VolMeter>>

The volume meter to draw on the actual Pixbuf, if requested.

audio_pix: RefCell<AudioPix>

The actual Pixbuf tray icon.

status_icon: StatusIcon

The gtk StatusIcon widget, used to register callbacks.

icon_size: Cell<i32>

The current icon size.

Methods

impl TrayIcon

fn new(prefs: &Prefs) -> Result<TrayIcon>

Constructor. audio_pix is initialized as empty GdkPixbuf, to save one iteration of png decoding (update_all() is triggered immediately on startup through tray_icon.connect_size_changed usually, otherwise we have to trigger it manually).

fn update_all(
&self,
prefs: &Prefs,
audio: &Audio,
m_size: Option<i32>
) -> Result<()>

Update the whole tray icon state.